Moved StickyHeaderWidth manipulation to main table js
